+++ xdot/__init__.py
@@ -1,0 +1,9 @@
+'''Visualize dot graphs via the xdot format.'''
+
+__all__ = ['dot', 'ui']
+
+from . import dot
+from . import ui
+from .ui import DotWidget, DotWindow
+
+__author__ = "Jose Fonseca et al"
