+++ xdot/ui/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['actions', 'animation', 'colours', 'elements', 'pen', 'window']
+
+from .window import DotWidget, DotWindow
