+++ xdot/ui/actions.py
@@ -1,0 +1,139 @@
+# Copyright 2008-2015 Jose Fonseca
+#
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+import gi
+gi.require_version('Gtk', '3.0')
+gi.require_version('PangoCairo', '1.0')
+
+from gi.repository import Gdk
+
+
+class DragAction(object):
+
+    def __init__(self, dot_widget):
+        self.dot_widget = dot_widget
+
+    def on_button_press(self, event):
+        self.startmousex = self.prevmousex = event.x
+        self.startmousey = self.prevmousey = event.y
+        self.start()
+
+    def on_motion_notify(self, event):
+        if event.is_hint:
+            window, x, y, state = event.window.get_device_position(event.device)
+        else:
+            x, y, state = event.x, event.y, event.state
+        deltax = self.prevmousex - x
+        deltay = self.prevmousey - y
+        self.drag(deltax, deltay)
+        self.prevmousex = x
+        self.prevmousey = y
+
+    def on_button_release(self, event):
+        self.stopmousex = event.x
+        self.stopmousey = event.y
+        self.stop()
+
+    def draw(self, cr):
+        pass
+
+    def start(self):
+        pass
+
+    def drag(self, deltax, deltay):
+        pass
+
+    def stop(self):
+        pass
+
+    def abort(self):
+        pass
+
+
+class NullAction(DragAction):
+
+    def on_motion_notify(self, event):
+        if event.is_hint:
+            window, x, y, state = event.window.get_device_position(event.device)
+        else:
+            x, y, state = event.x, event.y, event.state
+        dot_widget = self.dot_widget
+        item = dot_widget.get_url(x, y)
+        if item is None:
+            item = dot_widget.get_jump(x, y)
+        if item is not None:
+            dot_widget.get_window().set_cursor(Gdk.Cursor(Gdk.CursorType.HAND2))
+            dot_widget.set_highlight(item.highlight)
+        else:
+            dot_widget.get_window().set_cursor(None)
+            dot_widget.set_highlight(None)
+
+
+class PanAction(DragAction):
+
+    def start(self):
+        self.dot_widget.get_window().set_cursor(Gdk.Cursor(Gdk.CursorType.FLEUR))
+
+    def drag(self, deltax, deltay):
+        self.dot_widget.x += deltax / self.dot_widget.zoom_ratio
+        self.dot_widget.y += deltay / self.dot_widget.zoom_ratio
+        self.dot_widget.queue_draw()
+
+    def stop(self):
+        self.dot_widget.get_window().set_cursor(None)
+
+    abort = stop
+
+
+class ZoomAction(DragAction):
+
+    def drag(self, deltax, deltay):
+        self.dot_widget.zoom_ratio *= 1.005 ** (deltax + deltay)
+        self.dot_widget.zoom_to_fit_on_resize = False
+        self.dot_widget.queue_draw()
+
+    def stop(self):
+        self.dot_widget.queue_draw()
+
+
+class ZoomAreaAction(DragAction):
+
+    def drag(self, deltax, deltay):
+        self.dot_widget.queue_draw()
+
+    def draw(self, cr):
+        cr.save()
+        cr.set_source_rgba(.5, .5, 1.0, 0.25)
+        cr.rectangle(self.startmousex, self.startmousey,
+                     self.prevmousex - self.startmousex,
+                     self.prevmousey - self.startmousey)
+        cr.fill()
+        cr.set_source_rgba(.5, .5, 1.0, 1.0)
+        cr.set_line_width(1)
+        cr.rectangle(self.startmousex - .5, self.startmousey - .5,
+                     self.prevmousex - self.startmousex + 1,
+                     self.prevmousey - self.startmousey + 1)
+        cr.stroke()
+        cr.restore()
+
+    def stop(self):
+        x1, y1 = self.dot_widget.window2graph(self.startmousex,
+                                              self.startmousey)
+        x2, y2 = self.dot_widget.window2graph(self.stopmousex,
+                                              self.stopmousey)
+        self.dot_widget.zoom_to_area(x1, y1, x2, y2)
+
+    def abort(self):
+        self.dot_widget.queue_draw()
