+++ xdot/ui/pen.py
@@ -1,0 +1,54 @@
+# Copyright 2008-2015 Jose Fonseca
+#
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+BOLD = 1
+ITALIC = 2
+UNDERLINE = 4
+SUPERSCRIPT = 8
+SUBSCRIPT = 16
+STRIKE_THROUGH = 32
+OVERLINE = 64
+
+class Pen:
+    """Store pen attributes."""
+
+    def __init__(self):
+        # set default attributes
+        self.color = (0.0, 0.0, 0.0, 1.0)
+        self.fillcolor = (0.0, 0.0, 0.0, 1.0)
+        self.linewidth = 1.0
+        self.fontsize = 14.0
+        self.fontname = "Times-Roman"
+        self.bold = False
+        self.italic = False
+        self.underline = False
+        self.superscript = False
+        self.subscript = False
+        self.strikethrough = False
+        self.overline = False
+
+        self.dash = ()
+
+    def copy(self):
+        """Create a copy of this pen."""
+        pen = Pen()
+        pen.__dict__ = self.__dict__.copy()
+        return pen
+
+    def highlighted(self):
+        pen = self.copy()
+        pen.color = (1, 0, 0, 1)
+        pen.fillcolor = (1, .8, .8, 1)
+        return pen
